requirements.txt
transformers
pyyaml
psutil
langchain
